// src/types/DataRow.ts
export interface ResultDataRow {
    tid: number;
    angle: number;
    identifier: string;
    addIdentifier2: string;
    addIdentifier3: string;
    addIdentifier1: string;
    angleHighLimit: number;
    angleLowLimit: number;
    angleStatus: number;
    angleTarget: number;
    boltId: number;
    boltRevision: number;
    cycleNumber: number;
    dateTime: string;
    flags: number;
    generalStatus: number;
    jobCount: number;
    jobNumber: number;
    jobSize: number;
    opId: number;
    opRevision: number;
    programId: number | null;
    programNumber: number | null;
    spindleCount: number;
    spindleNumber: number;
    stationCode: number;
    stationState: number;
    toolAdr: string;
    toolDTO: {
      toolId: number;
      revision: number;
      toolName: string;
      toolAlias: string | null;
      insertId: number;
      ip: string;
      mac: string;
      deviceNumber: number;
      protocolID: number;
      stationCode: number;
      modelID: number;
      servoSerialNumber: string;
      toolSerialNumber: string;
      spindles: number;
      stationID: number;
      userID: number | null;
      insertDate: string;
      configStr: string;
      toolGroup: number;
      state: number;
    };
    toolProgramDTO: {
      programId: number;
      revision: number;
      toolId: number;
      insertDate: string;
      programNumber: number;
      programName: string;
      spindleNumber: number;
      strategy: number;
      torqueHighLimit: number;
      torqueLowLimit: number;
      torqueTarget: number;
      angleHighLimit: number;
      angleLowLimit: number;
      angleTarget: number;
      userId: number;
      state: number;
    };
    toolResultId: number;
    toolRevision: number | null;
    toolSerialNumber: string;
    torque: number;
    torqueHighLimit: number;
    torqueLowLimit: number;
    torqueStatus: number;
    torqueTarget: number;
    userId: number;
  }